rust: move ethereum crate functions to bitbox02-rust/hww/api/ethereum
No need for a separate crate - make it more consistent with how
bitcoin code is organized.
