+++ src/rust/Cargo.toml
@@ -18,7 +18,6 @@ members = [
     "bitbox02-rust-c",
     "bitbox02-rust",
     "util",
-    "apps/ethereum",
     "bitbox02-noise",
     "bitbox02",
     "bitbox02-sys",
