--- src/rust/apps/ethereum/Cargo.toml
@@ -1,25 +1,0 @@
-# Copyright 2020 Shift Cryptosecurity AG
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-[package]
-name = "ethereum"
-version = "0.1.0"
-authors = ["Shift Crypto AG <support@shiftcrypto.ch>"]
-edition = "2018"
-license = "Apache-2.0"
-
-[dependencies]
-util = {path = "../../util"}
-hex = { version = "0.4", default-features = false }
-sha3 = { version = "0.9.1", default-features = false }
