--- src/rust/apps/ethereum/src/lib.rs
@@ -1,26 +1,0 @@
-// Copyright 2020 Shift Cryptosecurity AG
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#![no_std]
-
-pub mod address;
-pub mod keypath;
-
-#[cfg(test)]
-#[macro_use]
-extern crate std;
-
-// for `format!`
-#[macro_use]
-extern crate alloc;
