+++ src/rust/bitbox02-rust-c/Cargo.toml
@@ -27,7 +27,6 @@ bitbox02-rust = { path = "../bitbox02-rust", optional = true }
 bitbox02 = { path = "../bitbox02", optional = true }
 bitbox02-noise = { path = "../bitbox02-noise", optional = true }
 util = { path = "../util" }
-ethereum = { path = "../apps/ethereum", optional = true }
 hex = { version = "0.4", default-features = false }
 sha2 = { version = "0.9.2", default-features = false, optional = true }
 
@@ -55,8 +54,6 @@ firmware = ["bitbox02-rust", "bitbox02", "bitbox02-noise", "sha2"]
 testing = ["bitbox02/testing"]
 
 app-ethereum = [
-  # these are dependencies
-  "ethereum",
   # enable this feature in the deps
   "bitbox02-rust/app-ethereum",
   "bitbox02/app-ethereum",
