+++ src/rust/bitbox02-rust/src/hww/api/ethereum/address.rs
@@ -1,3 +1,17 @@
+// Copyright 2021 Shift Crypto AG
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 use sha3::digest::Digest;
 
 use alloc::string::String;
@@ -38,7 +52,6 @@ pub fn from_pubkey(pubkey_uncompressed: &[u8; 65]) -> String {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use std::prelude::v1::*;
 
     #[test]
     fn test_from_pubkey_hash() {
